[refactor] Move #earlierThanTimeMills to TimeTravelUtil
