+++ paimon-core/src/main/java/org/apache/paimon/table/source/snapshot/TimeTravelUtil.java
@@ -23,18 +23,29 @@ import org.apache.paimon.Snapshot;
 import org.apache.paimon.schema.SchemaManager;
 import org.apache.paimon.schema.TableSchema;
 import org.apache.paimon.table.FileStoreTable;
+import org.apache.paimon.utils.ChangelogManager;
+import org.apache.paimon.utils.FunctionWithException;
 import org.apache.paimon.utils.SnapshotManager;
 import org.apache.paimon.utils.SnapshotNotExistException;
 import org.apache.paimon.utils.TagManager;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.annotation.Nullable;
+
+import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.apache.paimon.CoreOptions.SCAN_SNAPSHOT_ID;
 import static org.apache.paimon.utils.Preconditions.checkArgument;
+import static org.apache.paimon.utils.SnapshotManager.EARLIEST_SNAPSHOT_DEFAULT_RETRY_NUM;
 
 /** The util class of resolve snapshot from scan params for time travel. */
 public class TimeTravelUtil {
+
+    private static final Logger LOG = LoggerFactory.getLogger(TimeTravelUtil.class);
 
     private static final String[] SCAN_KEYS = {
         CoreOptions.SCAN_SNAPSHOT_ID.key(),
@@ -125,6 +136,106 @@ public class TimeTravelUtil {
         TagManager tagManager =
                 new TagManager(snapshotManager.fileIO(), snapshotManager.tablePath());
         return tagManager.getOrThrow(tagName).trimToSnapshot();
+    }
+
+    /**
+     * Returns the latest snapshot earlier than the timestamp mills. A non-existent snapshot may be
+     * returned if all snapshots are equal to or later than the timestamp mills.
+     */
+    public static @Nullable Long earlierThanTimeMills(
+            SnapshotManager snapshotManager,
+            ChangelogManager changelogManager,
+            long timestampMills,
+            boolean startFromChangelog) {
+        Long latest = snapshotManager.latestSnapshotId();
+        if (latest == null) {
+            return null;
+        }
+
+        Snapshot earliestSnapshot =
+                earliestSnapshot(snapshotManager, changelogManager, startFromChangelog, latest);
+        if (earliestSnapshot == null) {
+            return latest - 1;
+        }
+
+        if (earliestSnapshot.timeMillis() >= timestampMills) {
+            return earliestSnapshot.id() - 1;
+        }
+
+        long earliest = earliestSnapshot.id();
+        while (earliest < latest) {
+            long mid = (earliest + latest + 1) / 2;
+            Snapshot snapshot =
+                    startFromChangelog
+                            ? changelogOrSnapshot(snapshotManager, changelogManager, mid)
+                            : snapshotManager.snapshot(mid);
+            if (snapshot.timeMillis() < timestampMills) {
+                earliest = mid;
+            } else {
+                latest = mid - 1;
+            }
+        }
+        return earliest;
+    }
+
+    private static @Nullable Snapshot earliestSnapshot(
+            SnapshotManager snapshotManager,
+            ChangelogManager changelogManager,
+            boolean includeChangelog,
+            @Nullable Long stopSnapshotId) {
+        Long snapshotId = null;
+        if (includeChangelog) {
+            snapshotId = changelogManager.earliestLongLivedChangelogId();
+        }
+        if (snapshotId == null) {
+            snapshotId = snapshotManager.earliestSnapshotId();
+        }
+        if (snapshotId == null) {
+            return null;
+        }
+
+        if (stopSnapshotId == null) {
+            stopSnapshotId = snapshotId + EARLIEST_SNAPSHOT_DEFAULT_RETRY_NUM;
+        }
+
+        FunctionWithException<Long, Snapshot, FileNotFoundException> snapshotFunction =
+                includeChangelog
+                        ? s -> tryGetChangelogOrSnapshot(snapshotManager, changelogManager, s)
+                        : snapshotManager::tryGetSnapshot;
+
+        do {
+            try {
+                return snapshotFunction.apply(snapshotId);
+            } catch (FileNotFoundException e) {
+                snapshotId++;
+                if (snapshotId > stopSnapshotId) {
+                    return null;
+                }
+                LOG.warn(
+                        "The earliest snapshot or changelog was once identified but disappeared. "
+                                + "It might have been expired by other jobs operating on this table. "
+                                + "Searching for the second earliest snapshot or changelog instead. ");
+            }
+        } while (true);
+    }
+
+    private static Snapshot tryGetChangelogOrSnapshot(
+            SnapshotManager snapshotManager, ChangelogManager changelogManager, long snapshotId)
+            throws FileNotFoundException {
+        if (changelogManager.longLivedChangelogExists(snapshotId)) {
+            return changelogManager.tryGetChangelog(snapshotId);
+        } else {
+            return snapshotManager.tryGetSnapshot(snapshotId);
+        }
+    }
+
+    private static Snapshot changelogOrSnapshot(
+            SnapshotManager snapshotManager, ChangelogManager changelogManager, long snapshotId) {
+        if (changelogManager.longLivedChangelogExists(snapshotId)) {
+            return changelogManager.changelog(snapshotId);
+        } else {
+            return snapshotManager.snapshot(snapshotId);
+        }
     }
 
     public static void checkRescaleBucketForIncrementalTagQuery(
